Use Lua implementation of conversion to string
Previously we had an inconsistency where we would use Lua's definition of
conversion to bool but cty's definition of conversion to string.

Here we unify this by using Lua's definition of conversion to string
also, which means that fewer Lua types can be used when a string is
expected but creates better consistency with how Lua behaves in other
scenarios.